move function to new py file
