+++ src/plot_binary.py
@@ -1,0 +1,59 @@
+import pandas as pd
+
+def plot_binary(
+        input_df: pd.DataFrame,
+        list_of_variables: list,
+        plot_kind: str,
+        label_y_offset: int,
+        label_fontsize: int,
+        figsize: tuple = (10, 6),
+        output_path: str = None,
+        super_title: str = "Distribution of Binary Variables",
+        super_title_font: int = 14) -> None:
+    """Plot the distribution of binary variables in a DataFrame, save the plot, and display it.
+
+    This function will construct a set of subplots (either a bar plot or a pie chart)
+    for each of the binary variables specified in `list_of_variables`. The function will
+    save as well as display the entire figure. 
+
+    Parameters
+    ----------
+    input_df : pd.DataFrame
+        The DataFrame that contains the binary variables.
+
+    list_of_variables : list
+        List of binary variables (column names) to be plotted.
+
+    figsize: tuple[int, int]:
+        The width and height of the figure size in a tuple.
+
+    plot_kind : {'count', 'pie'}
+        Type of plot to be generated, a count plot or a pie chart.
+
+    label_y_offset : int
+        Y-axis offset for label positioning.
+
+    label_fontsize : int
+        Font size for axis labels.
+
+    output_path : str, optional
+        Path to save the plot. Defaults to the current working directory.
+
+    super_title : str, optional
+        Super title for the entire plot. Default is "Distribution of Binary Variables".
+
+    super_title_font : int, optional
+        Font size for the super title. Default is 14.
+
+    Returns
+    -------
+    None
+        
+
+    Examples
+    --------
+    binary = ["binary_var1", "binary_var2"]
+    plot_binary(df, binary, "h", 30, 10, "count", 10, 10, "sample.png")
+    """
+    
+    pass # Implementation to be added
